requirements.txt
# Optional dependencies for full pipeline
pymupdf>=1.24.0
pillow>=10.0.0
opencv-python>=4.9.0
numpy>=1.26.0
pydantic>=2.7.0
pyyaml>=6.0.1
requests>=2.31.0
# milvus
pymilvus>=2.4.0
# faiss (CPU)
faiss-cpu>=1.7.4
